Fix windows build
work-items: #50884
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C53547

+++ development/libraries/utilities/src/ProgressManager.C
@@ -25,7 +25,7 @@ std::string ProgressManager::report (const std::string & message)
 
    WallTime::Duration executionTime = clockTime - m_timeStarted;
 
-   int len = message.size ();
+   size_t len = message.size ();
 
    ostringstream buf;
    buf << message;
